Begin service-split refactor
Now that I'm looking at how this'll actually work externally, I'm
questioning my encapsulation-over-inheritance approach: I would love for

```js
var MapboxClient = require('mapbox/services/geocoder');
```

To export an object with a `.geocoder.forward` method instead
of `.forward` in the top namespace, so that documentation doesn't have
to deal with the alternative possible name.
