+++ lib/make_query.js
@@ -1,0 +1,10 @@
+'use strict';
+
+var qs = require('querystring');
+
+function makeQuery(accessToken, options) {
+  options.access_token = accessToken;
+  return '?' + qs.stringify(options);
+}
+
+module.exports = makeQuery;
